+++ sql/stress-fuseki.sql
@@ -2,7 +2,8 @@ CREATE SERVER fuseki
 FOREIGN DATA WRAPPER rdf_fdw 
 OPTIONS (
   endpoint   'http://fuseki:3030/dt/sparql',
-  update_url 'http://fuseki:3030/dt/update');
+  update_url 'http://fuseki:3030/dt/update',
+  batch_size '10000');
 
 CREATE FOREIGN TABLE ft (
   subject   rdfnode OPTIONS (variable '?s'),
@@ -22,29 +23,29 @@ SERVER fuseki OPTIONS (
 CREATE USER MAPPING FOR postgres
 SERVER fuseki OPTIONS (user 'admin', password 'secret');
 
-/* inserting a large literal (100k characters) */
+/* inserting a large literal (one million characters) */
 INSERT INTO ft (subject, predicate, object)
-VALUES  ('<https://www.uni-muenster.de>', '<http://www.w3.org/2000/01/rdf-schema#label>', repeat('x', 100000)::rdfnode);
+VALUES  ('<https://www.uni-muenster.de>', '<http://www.w3.org/2000/01/rdf-schema#label>', repeat('x', 1000000)::rdfnode);
 
 SELECT subject, predicate, sparql.strlen(object) 
 FROM ft;
 
-SELECT subject, predicate, sparql.strlen(object) 
-FROM ft;
-
-/* updating the large literal (100k characters) */
+/* updating the large literal (one million characters) */
 UPDATE ft SET
-  object = repeat('y', 100000)::rdfnode
+  object = repeat('y', 1000000)::rdfnode
 WHERE subject = '<https://www.uni-muenster.de>'
   AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+
+SELECT subject, predicate, sparql.strlen(object) 
+FROM ft;
 
 /* deleting triple with large literal */
 DELETE FROM ft;
 
-/* inserting a 100k triples */
+/* inserting a one million triples */
 INSERT INTO ft (subject, predicate, object)
 SELECT '<https://www.uni-muenster.de>', '<http://www.w3.org/2000/01/rdf-schema#label>', i::rdfnode 
-FROM generate_series(1,100000) AS g(i);
+FROM generate_series(1,1000000) AS g(i);
 
 SELECT * FROM ft
 LIMIT 5;